ACD-638: Remove hearings feature flag (#2891)
* Remove hearings FF

* Address review comments
