+++ .k8s/live/dev/deployment.yaml
@@ -65,10 +65,6 @@ spec:
               value: enabled
             - name: DISPLAY_RAW_RESPONSES
               value: enabled
-            - name: DEFENDANTS_SEARCH
-              value: 'true'
-            - name: HEARING
-              value: 'true'
             - name: COURT_DATA_ADAPTOR_API_UID
               valueFrom:
                 secretKeyRef:
