+++ .k8s/live/uat/deployment.yaml
@@ -63,10 +63,6 @@ spec:
               value: 'UA-131160087-17'
             - name: RAILS_LOG_TO_STDOUT
               value: enabled
-            - name: DEFENDANTS_SEARCH
-              value: 'true'
-            - name: HEARING
-              value: 'true'
             - name: DISPLAY_RAW_RESPONSES
               value: enabled
             - name: COURT_DATA_ADAPTOR_API_UID
