+++ app/helpers/hearing_paginator.rb
@@ -90,14 +90,8 @@ class HearingPaginator
   def sorted_hearing_items
     @prosecution_case.hearings_sort_column = @column
     @prosecution_case.hearings_sort_direction = @direction
-    if FeatureFlag.enabled?(:hearing)
-      @prosecution_case.sorted_hearing_summaries_with_day.map do |hearing|
-        PageItem.new(hearing.id, hearing.day)
-      end
-    else
-      @prosecution_case.sorted_hearings_with_day.map do |hearing|
-        PageItem.new(hearing.id, hearing.day)
-      end
+    @prosecution_case.sorted_hearing_summaries_with_day.map do |hearing|
+      PageItem.new(hearing.id, hearing.day)
     end
   end
 end
