--- app/views/hearings/_attendees_v2.html.haml
@@ -1,32 +1,0 @@
-%div{ :class => "govuk-!-margin-bottom-9" }
-  %h2.govuk-heading-l
-    = t('hearings.show.attendees.heading')
-  .govuk-heading-s{ :class => "govuk-!-margin-bottom-1" }
-    = t('hearings.show.attendees.defendants')
-  %p.govuk-body{ :class => "govuk-!-margin-bottom-2", :id => "defendants" }
-    - unless hearing_details && hearing_details.prosecution_cases&.any?
-      = t('generic.not_available')
-    - else
-      - cases = hearing_details.prosecution_cases
-      = safe_join(cases.map { |cs| cs.defendants.map {|df| "#{df.defendant_details.person_details.first_name} #{df.defendant_details.person_details.last_name}"}}, tag.br)
-  .govuk-heading-s{ :class => "govuk-!-margin-bottom-1" }
-    = t('hearings.show.attendees.defence')
-  %p.govuk-body{ :class => "govuk-!-margin-bottom-2", :id => "defence" }
-    - unless hearing_details && hearing_details.defence_counsels&.any?
-      = t('generic.not_available')
-    - else
-      = safe_join([hearing.defence_counsels_list], tag.br)
-  .govuk-heading-s{ :class => "govuk-!-margin-bottom-1" }
-    = t('hearings.show.attendees.prosecution')
-  %p.govuk-body{ :class => "govuk-!-margin-bottom-2", :id => "prosecution" }
-    - unless hearing_details && hearing_details.prosecution_counsels&.any?
-      = t('generic.not_available')
-    - else
-      = hearing.prosecution_counsels_list
-  .govuk-heading-s{ :class => "govuk-!-margin-bottom-1" }
-    = t('hearings.show.attendees.judges')
-  %p.govuk-body{ :class => "govuk-!-margin-bottom-2", :id => "judges" }
-    - unless hearing_details && hearing_details.judiciary&.any?
-      = t('generic.not_available')
-    - else
-      = hearing.judiciary_list
