--- app/views/hearings/_court_applications_v2.html.haml
@@ -1,37 +1,0 @@
-:ruby
-  not_available = t('generic.not_available')
-  respondent_synonyms = t('hearings.show.court_applications.respondent_synonyms')
-  applicant_synonym = t('hearings.show.court_applications.applicant_synonym')
-
-%h2.govuk-heading-l{ :class => "govuk-!-margin-bottom-4" }
-  = t('hearings.show.court_applications.heading')
-- if hearing && hearing.court_applications.present?
-  .govuk-accordion{ :data => {:module => "govuk-accordion"}, :id => "accordion-" + hearing.id }
-    - hearing.court_applications.each do |court_application|
-      .govuk-accordion__section
-        .govuk-accordion__section-header
-          %h3.govuk-accordion__section-heading
-            %span.govuk-accordion__section-button
-              = court_application&.type&.description || not_available
-          .govuk-accordion__section-summary.govuk-body
-            = l(court_application.received_date&.to_date, :format => :long)
-        .govuk-accordion__section-content.govuk-details__text
-          - key_class = {}
-          - value_class = {class: "vcd-capitalize-first-letter"}
-          %dl.govuk-summary-list
-            = govuk_summary_list_entry(t('hearings.show.court_applications.received_date'), l(court_application.received_date&.to_date, :format => :long), key_class, value_class)
-            = govuk_summary_list_entry(t('hearings.show.court_applications.type'), court_application&.type&.description || not_available, key_class, value_class)
-            = govuk_summary_list_entry(t('hearings.show.court_applications.code'), court_application&.type&.code || not_available, key_class, value_class)
-            = govuk_summary_list_entry(t('hearings.show.court_applications.legislation'), court_application&.type&.legislation  || not_available, key_class, value_class)
-            - court_application.judicial_results&.each do |result|
-              = govuk_summary_list_entry(t('hearings.show.court_applications.result_codes'), result&.cjs_code || not_available, key_class, value_class)
-              - if result&.text.present?
-                = govuk_summary_list_entry(t('hearings.show.court_applications.result_text'), transform_and_sanitize(result&.text), key_class, value_class)
-              - else
-                = govuk_summary_list_entry(t('hearings.show.court_applications.result_text'), not_available, key_class, value_class)
-            = govuk_summary_list_entry(respondent_synonyms, safe_join(court_application.respondents.map { |rs| rs.synonym }, tag.br), key_class, value_class)
-            = govuk_summary_list_entry(applicant_synonym, court_application.type.applicant_appellant_flag ? 'Applicant' : '', key_class, value_class)
-
-- else
-  %p.govuk-body
-    = t('hearings.show.court_applications.none')
