--- app/views/hearings/_details.haml
@@ -1,15 +1,0 @@
-%div{ :class => "govuk-!-margin-bottom-9" }
-  .govuk-heading-l
-    = t('generic.detail')
-  .govuk-heading-s{ :class => "govuk-!-margin-bottom-1" }
-    = t('hearings.show.hearing_type')
-  %p.govuk-body{ :class => "govuk-!-margin-bottom-2" }
-    = hearing&.hearing_type || t('generic.not_available')
-  .govuk-heading-s{ :class => "govuk-!-margin-bottom-1" }
-    = t('hearings.show.court')
-  %p.govuk-body{ :class => "govuk-!-margin-bottom-2" }
-    = hearing&.court_name || t('generic.not_available')
-  .govuk-heading-s{ :class => "govuk-!-margin-bottom-1" }
-    = t('hearings.show.time_listed')
-  %p.govuk-body{ :class => "govuk-!-margin-bottom-2" }
-    = hearing_day&.strftime('%H:%M') || t('generic.not_available')
