+++ app/views/hearings/_details.html.haml
