--- app/views/hearings/_hearing_events_v2.html.haml
@@ -1,19 +1,0 @@
-.govuk-heading-l
-  = t('hearings.show.events.title')
-%table.govuk-table{ :class => "govuk-!-margin-bottom-9" }
-  %caption.govuk-table__caption.govuk-visually-hidden
-    = t('hearings.show.events.caption')
-  %thead.govuk-table__head
-    %tr.govuk-table__row
-      %th.govuk-table__header{ scope: 'col' }= t('hearings.show.events.time')
-      %th.govuk-table__header{ scope: 'col' }= t('hearings.show.events.event')
-  %tbody.govuk-table__body
-    - hearing_events.events.each do |event|
-      %tr.govuk-table__row
-        %td.govuk-table__cell
-          = event.event_time.to_datetime.strftime('%H:%M')
-        %td.govuk-table__cell.hearing-events-table__cell--overflow-wrap
-          = event.recorded_label
-          - if event&.note.present?
-            %div{ class: 'govuk-!-padding-top-1' }
-              = transform_and_sanitize(event.note)
