--- app/views/hearings/_hearing_result_v2.html.haml
@@ -1,5 +1,0 @@
-%div{ :class => "govuk-!-margin-bottom-9" }
-  .govuk-heading-l
-    = t('hearings.show.result')
-  %p.govuk-body
-    = sanitize(hearing.cracked_ineffective_trial.description_sentence)
