+++ app/views/hearings/_raw_response.html.haml
@@ -1,25 +1,11 @@
-%div{ class: "govuk-!-padding-bottom-4" }
-= govuk_detail(t("generic.view_raw_data")) do
-  %h3= "Hearings"
+%div{ class: 'govuk-!-padding-bottom-4' }
+= govuk_detail(t('generic.view_raw_data')) do
+  %h3= 'Hearing Summaries'
   %pre
-    = JSON.pretty_generate(hearing&.attributes)
-  %h3= "Court applications"
-  - hearing&.court_applications&.each.with_index do |court_application, caidx|
-    %pre
-      = JSON.pretty_generate(hearing&.court_applications&.map(&:attributes))
-      %h4 Court application type for court application #{caidx}
-      %pre
-        = JSON.pretty_generate(court_application.type&.attributes)
-      %h4 Judicial results for court application #{caidx}
-      = "none" if court_application.judicial_results.blank?
-      - Array(court_application.judicial_results).each.with_index do |result, ridx|
-        %pre
-          = JSON.pretty_generate(result.attributes)
-      %h4 Respondents for court application #{caidx}
-      = "none" if court_application.respondents.blank?
-      - Array(court_application.respondents).each.with_index do |respondent, ridx|
-        %pre
-          = JSON.pretty_generate(respondent.attributes)
-  %h3= "Hearing events"
+    = JSON.pretty_generate(hearing_summary.as_json)
+  %h3= 'Hearing'
   %pre
-    = JSON.pretty_generate(hearing&.hearing_events&.map(&:attributes))
+    = JSON.pretty_generate(hearing.as_json)
+  %h3= 'Hearing events'
+  %pre
+    = JSON.pretty_generate(hearing_events.as_json)
