--- app/views/prosecution_cases/_hearing_summaries.html.haml
@@ -1,23 +1,0 @@
-%h2.govuk-heading-l{ :class => "govuk-!-margin-bottom-4" }
-  = t('search.result.hearing.heading')
-%table.govuk-table.app-sortable-table{ :class => "govuk-!-margin-bottom-9" }
-  %caption.govuk-table__caption.govuk-visually-hidden
-    = t('search.result.hearing.caption')
-  %thead.govuk-table__head
-    %tr.govuk-table__row
-      %th.govuk-table__header{ scope: 'col' }
-        = hearings_sorter_link(prosecution_case, 'date')
-      %th.govuk-table__header{ scope: 'col' }
-        = hearings_sorter_link(prosecution_case, 'type')
-      %th.govuk-table__header{ scope: 'col' }
-        = hearings_sorter_link(prosecution_case, 'provider')
-  %tbody.govuk-table__body
-    - prosecution_case.sorted_hearings_with_day.each_with_index do |hearing, idx|
-      %tr.govuk-table__row
-        %td.govuk-table__cell
-          = link_to hearing.day.strftime('%d/%m/%Y'), hearing_path(id: hearing.id, urn: prosecution_case.prosecution_case_reference, column: prosecution_case.hearings_sort_column,
-          direction: prosecution_case.hearings_sort_direction, page: idx), class: 'govuk-link'
-        %td.govuk-table__cell
-          = hearing.hearing_type
-        %td.govuk-table__cell
-          = hearing ? hearing.provider_list : t('generic.not_available')
