--- app/views/prosecution_cases/_raw_response.html.haml
@@ -1,44 +1,0 @@
-%div{ class: 'govuk-!-padding-bottom-4' }
-= govuk_detail(t('generic.view_raw_data')) do
-  %h3= t('generic.prosecution_case')
-  %pre
-    = JSON.pretty_generate(prosecution_case.as_json)
-
-  %h3= 'Defendants'
-  %pre
-    - prosecution_case.defendants.each.with_index do |defendant, idx|
-      %h4 defendant #{idx}
-      = JSON.pretty_generate(defendant.attributes)
-      %h4 Offences for defendant #{idx}
-      - defendant.offences.each do |offence|
-        = JSON.pretty_generate(offence.attributes)
-
-  %h3= 'Hearing summaries'
-  - prosecution_case.hearing_summaries.each do |hearing_summary|
-    %pre
-      = JSON.pretty_generate(hearing_summary.attributes)
-
-  %h3= 'Hearings'
-  - prosecution_case.hearings.each.with_index do |hearing, hidx|
-    %h4 hearing #{hidx}
-    %pre
-      = JSON.pretty_generate(hearing.attributes)
-
-    %h4 Hearing events for hearing #{hidx}
-    = 'none' if hearing.hearing_events.blank?
-    - Array(hearing.hearing_events).each.with_index do |event, eidx|
-      %pre
-        = JSON.pretty_generate(event.attributes)
-
-    %h4 Providers for hearing #{hidx}
-    = 'none' if hearing.providers.blank?
-    - Array(hearing.providers).each.with_index do |event, pidx|
-      %pre
-        = JSON.pretty_generate(event.attributes)
-
-    %h4 Cracked ineffective details for hearing #{hidx}
-    - if hearing.cracked_ineffective_trial.nil?
-      none
-    - else
-      %pre
-        = JSON.pretty_generate(hearing.cracked_ineffective_trial.attributes)
