+++ spec/decorators/cd_api/cracked_ineffective_trial_decorator_spec.rb
@@ -13,10 +13,6 @@ RSpec.describe CdApi::CrackedIneffectiveTrialDecorator, type: :decorator do
     end
   end
 
-  before do
-    allow(FeatureFlag).to receive(:enabled?).with(:hearing).and_return(true)
-  end
-
   it_behaves_like 'a base decorator' do
     let(:object) { cracked_ineffective_trial }
   end
