+++ spec/views/hearings/show.html.haml_spec.rb
@@ -1,77 +1,58 @@
 # frozen_string_literal: true
 
-RSpec.describe 'hearings/show', type: :view do
+RSpec.describe 'hearings/show', :stub_v2_hearing_data, :stub_v2_hearing_summary, type: :view do
   subject(:render_view) { render }
 
-  # rubocop: disable RSpec/VerifiedDoubles
-  # NOTE: an instance double would require more setup
-  # and we are only testing partials are rendered.
+  let(:case_reference) { 'TEST12345' }
   let(:prosecution_case) do
-    double(ProsecutionCaseDecorator,
-           hearings: [hearing],
-           prosecution_case_reference: 'ACASEURN',
-           sorted_hearings_with_day: [hearing],
-           hearings_sort_column: 'date',
-           hearings_sort_direction: 'asc')
+    build(:case_summary, :with_hearing_summaries,
+          prosecution_case_reference: case_reference)
   end
-  # rubocop: enable RSpec/VerifiedDoubles
-
-  let(:hearing) { CourtDataAdaptor::Resource::Hearing.new(hearing_events: []) }
-  let(:decorated_hearing) { view.decorate(hearing) }
-  let(:hearing_day) { Date.parse('2021-01-17T10:30:00.000Z') }
+  let(:decorated_prosecution_case) { view.decorate(prosecution_case, CdApi::CaseSummaryDecorator) }
+  let(:hearing_id) { '844a6542-ffcb-4cd0-94ce-fda3ffc3081b' }
+  let(:hearing_day) { Date.parse('2019-10-23T10:30:00.000Z') }
+  let(:hearing_events) do
+    CdApi::HearingEvents.find(hearing_id,
+                              params: { date: hearing_day.strftime('%F') })
+  end
+  let(:hearing) do
+    view.decorate(CdApi::Hearing.find(hearing_id,
+                                      params: { date: hearing_day.strftime('%F') }), CdApi::HearingDecorator)
+  end
   let(:paginator) do
-    HearingPaginator.new(prosecution_case, column: 'date', direction: 'asc', page: '0')
+    HearingPaginator.new(decorated_prosecution_case, column: 'date', direction: 'asc', page: '0')
   end
 
   before do
-    allow(view).to receive(:govuk_page_title).and_return 'A Gov uk page title'
-    allow(prosecution_case).to receive(:hearings_sort_column=)
-    allow(prosecution_case).to receive(:hearings_sort_direction=)
-    assign(:hearing, decorated_hearing)
+    allow(view).to receive(:govuk_page_title).and_return 'Hearings Page'
+
+    assign(:hearing, hearing)
     assign(:hearing_day, hearing_day)
     assign(:paginator, paginator)
+    assign(:prosecution_case, decorated_prosecution_case)
   end
 
-  it { is_expected.to have_content('A Gov uk page title') }
-  it { is_expected.to render_template(:_pagination) }
-  it { is_expected.to render_template(:_details) }
-  it { is_expected.to render_template(:_attendees) }
-  it { is_expected.to render_template(:_hearing_events) }
-  it { is_expected.to render_template(:_court_applications) }
-
-  context 'when hearing_events is missing' do
-    let(:hearing) { CourtDataAdaptor::Resource::Hearing.new }
-
-    it { is_expected.not_to render_template(:_hearing_events) }
+  context 'when viewing hearing details' do
+    it 'displays the partial' do
+      is_expected.to render_template('hearings/_details')
+    end
   end
 
-  context 'with cracked_ineffective_trial' do
-    let(:cracked_ineffective_trial) do
-      CourtDataAdaptor::Resource::CrackedIneffectiveTrial
-        .new(id: 'a-uuid',
-             type: 'Ineffective',
-             description: 'Another case over-ran')
+  context 'when viewing attendees' do
+    it 'displays the partial' do
+      is_expected.to render_template('hearings/_attendees')
     end
-
-    before do
-      allow(hearing)
-        .to receive(:cracked_ineffective_trial)
-        .and_return(cracked_ineffective_trial)
-    end
-
-    it { is_expected.to have_css('div.govuk-heading-s', text: /Result/) }
-    it { is_expected.to have_css('p.govuk-body', text: /Ineffective: Another case over-ran/) }
   end
 
-  context 'without cracked_ineffective_trial' do
-    let(:cracked_ineffective_trial) { nil }
-
-    before do
-      allow(hearing)
-        .to receive(:cracked_ineffective_trial)
-        .and_return(cracked_ineffective_trial)
+  context 'when viewing results' do
+    it 'displays the partial' do
+      is_expected.to render_template('hearings/_hearing_result')
     end
+  end
 
-    it { is_expected.to have_no_css('h2.govuk-heading-l', text: /Result/) }
+  context 'when viewing court applications' do
+    it 'displays the partial' do
+      is_expected.to render_template('hearings/_court_applications')
+    end
   end
 end
